+++ sdk/program/src/sysvar/epoch_rewards.rs
@@ -130,10 +130,17 @@
 //! ```
 
 pub use crate::epoch_rewards::EpochRewards;
-use crate::{impl_sysvar_get, program_error::ProgramError, sysvar::Sysvar};
+use {
+    crate::{
+        impl_sysvar_get,
+        program_error::ProgramError,
+        sysvar::{GettableSysvar, Sysvar},
+    },
+    num_traits::ToPrimitive,
+};
 
 crate::declare_sysvar_id!("SysvarEpochRewards1111111111111111111111111", EpochRewards);
 
 impl Sysvar for EpochRewards {
-    impl_sysvar_get!(sol_get_epoch_rewards_sysvar);
+    impl_sysvar_get!(EpochRewards);
 }
