+++ sdk/program/src/sysvar/epoch_schedule.rs
@@ -126,10 +126,17 @@
 //! # Ok::<(), anyhow::Error>(())
 //! ```
 pub use crate::epoch_schedule::EpochSchedule;
-use crate::{impl_sysvar_get, program_error::ProgramError, sysvar::Sysvar};
+use {
+    crate::{
+        impl_sysvar_get,
+        program_error::ProgramError,
+        sysvar::{GettableSysvar, Sysvar},
+    },
+    num_traits::ToPrimitive,
+};
 
 crate::declare_sysvar_id!("SysvarEpochSchedu1e111111111111111111111111", EpochSchedule);
 
 impl Sysvar for EpochSchedule {
-    impl_sysvar_get!(sol_get_epoch_schedule_sysvar);
+    impl_sysvar_get!(EpochSchedule);
 }
