+++ sdk/program/src/sysvar/last_restart_slot.rs
@@ -40,7 +40,14 @@
 //!
 
 pub use crate::last_restart_slot::LastRestartSlot;
-use crate::{impl_sysvar_get, program_error::ProgramError, sysvar::Sysvar};
+use {
+    crate::{
+        impl_sysvar_get,
+        program_error::ProgramError,
+        sysvar::{GettableSysvar, Sysvar},
+    },
+    num_traits::ToPrimitive,
+};
 
 crate::declare_sysvar_id!(
     "SysvarLastRestartS1ot1111111111111111111111",
@@ -48,5 +55,5 @@ crate::declare_sysvar_id!(
 );
 
 impl Sysvar for LastRestartSlot {
-    impl_sysvar_get!(sol_get_last_restart_slot);
+    impl_sysvar_get!(LastRestartSlot);
 }
