+++ sdk/program/src/sysvar/rent.rs
@@ -127,10 +127,17 @@
 //! # Ok::<(), anyhow::Error>(())
 //! ```
 pub use crate::rent::Rent;
-use crate::{impl_sysvar_get, program_error::ProgramError, sysvar::Sysvar};
+use {
+    crate::{
+        impl_sysvar_get,
+        program_error::ProgramError,
+        sysvar::{GettableSysvar, Sysvar},
+    },
+    num_traits::ToPrimitive,
+};
 
 crate::declare_sysvar_id!("SysvarRent111111111111111111111111111111111", Rent);
 
 impl Sysvar for Rent {
-    impl_sysvar_get!(sol_get_rent_sysvar);
+    impl_sysvar_get!(Rent);
 }
